# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Budget Time-Machine production planning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /room-context:
    get:
      operationId: getRoomContext
      tags: [rooms]
      summary: Resolve the current browser room and access context
      responses:
        "200":
          description: Room and optional signed-in producer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomContextResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /rooms:
    get:
      operationId: listProductionRooms
      tags: [rooms]
      summary: List rooms owned by or shared with the producer
      responses:
        "200":
          description: Accessible production rooms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListRoomsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createProductionRoom
      tags: [rooms]
      summary: Create a named team production room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomInput"
      responses:
        "201":
          description: Created production room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /rooms/{roomId}/claim:
    post:
      operationId: claimProductionRoom
      tags: [rooms]
      summary: Claim the current anonymous browser room
      parameters:
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimRoomInput"
      responses:
        "200":
          description: Claimed production room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /rooms/{roomId}/invites:
    post:
      operationId: createRoomInvite
      tags: [rooms]
      summary: Create an email-bound room invite
      parameters:
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInviteInput"
      responses:
        "201":
          description: One-time invite token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomInviteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /room-invites/{token}/accept:
    post:
      operationId: acceptRoomInvite
      tags: [rooms]
      summary: Accept a room invite for the signed-in producer
      parameters:
        - $ref: "#/components/parameters/InviteToken"
      responses:
        "200":
          description: Joined room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptInviteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /rooms/{roomId}/members:
    get:
      operationId: listRoomMembers
      tags: [rooms]
      summary: List collaborators in a production room
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: Room members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListMembersResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  parameters:
...
    RoomId:
      name: roomId
      in: path
      required: true
      schema: { type: string, pattern: "^[a-zA-Z0-9_-]{8,64}$" }
    InviteToken:
      name: token
      in: path
      required: true
      schema: { type: string, minLength: 16 }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Confirmation required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ProducerUser:
      type: object
      properties:
        userId: { type: string }
        email: { type: string, format: email }
        displayName: { type: string }
      required: [userId, email, displayName]
...
    Room:
      type: object
      properties:
        roomId: { type: string }
        name: { type: string }
        role: { type: string, enum: [owner, editor, viewer, anonymous] }
        visibility: { type: string, enum: [private, team] }
        anonymous: { type: boolean }
        hasProduction: { type: boolean }
        updatedAt: { type: string, format: date-time }
      required: [roomId, name, role, visibility]
    RoomContextResponse:
      type: object
      properties:
        room:
          $ref: "#/components/schemas/Room"
        user:
          oneOf:
            - $ref: "#/components/schemas/ProducerUser"
            - type: "null"
      required: [room, user]
    ListRoomsResponse:
      type: object
      properties:
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/Room"
      required: [rooms]
    RoomResponse:
      type: object
      properties:
        room:
          $ref: "#/components/schemas/Room"
      required: [room]
    CreateRoomInput:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
      required: [name]
    ClaimRoomInput:
      type: object
      properties:
        name: { type: ["string", "null"], maxLength: 120 }
    RoomInviteInput:
      type: object
      properties:
        email: { type: string, format: email, maxLength: 255 }
        role: { type: string, enum: [editor, viewer] }
      required: [email]
    RoomInviteResponse:
      type: object
      properties:
        token: { type: string }
        expiresInDays: { type: number }
      required: [token, expiresInDays]
    AcceptInviteResponse:
      type: object
      properties:
        roomId: { type: string }
      required: [roomId]
    RoomMember:
      type: object
      properties:
        userId: { type: string }
        email: { type: string, format: email }
        displayName: { type: string }
        role: { type: string, enum: [owner, editor, viewer] }
      required: [userId, email, displayName, role]
    ListMembersResponse:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/RoomMember"
      required: [members]
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...