# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Budget Time-Machine production planning API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomRole } from './roomRole';
import type { RoomVisibility } from './roomVisibility';

export interface Room {
  roomId: string;
  name: string;
  role: RoomRole;
  visibility: RoomVisibility;
  anonymous?: boolean;
  hasProduction?: boolean;
  updatedAt?: Date;
}