# lib/api-zod/src/generated/types/roomMember.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Budget Time-Machine production planning API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomMemberRole } from './roomMemberRole';

export interface RoomMember {
  userId: string;
  email: string;
  displayName: string;
  role: RoomMemberRole;
}